# fix: improve hashBlocks() in as-sha256