--- a/packages/as-sha256/build/untouched.wat
+++ b/packages/as-sha256/build/untouched.wat
@@ -52,6 +52,7 @@
  (global $assembly/index/outputPtr (mut i32) (i32.const 0))
  (global $assembly/index/mLength (mut i32) (i32.const 0))
  (global $assembly/index/bytesHashed (mut i32) (i32.const 0))
+ (global $assembly/index/tmpW (mut i32) (i32.const 0))
  (global $~lib/memory/__heap_base i32 (i32.const 1012))
  (memory $0 1)
  (data $0 (i32.const 12) "\1c\01\00\00\00\00\00\00\00\00\00\00\01\00\00\00\00\01\00\00\98/\8aB\91D7q\cf\fb\c0\b5\a5\db\b5\e9[\c2V9\f1\11\f1Y\a4\82?\92\d5^\1c\ab\98\aa\07\d8\01[\83\12\be\851$\c3}\0cUt]\ber\fe\b1\de\80\a7\06\dc\9bt\f1\9b\c1\c1i\9b\e4\86G\be\ef\c6\9d\c1\0f\cc\a1\0c$o,\e9-\aa\84tJ\dc\a9\b0\\\da\88\f9vRQ>\98m\c61\a8\c8\'\03\b0\c7\7fY\bf\f3\0b\e0\c6G\91\a7\d5Qc\ca\06g))\14\85\n\b7\'8!\1b.\fcm,M\13\r8STs\ne\bb\njv.\c9\c2\81\85,r\92\a1\e8\bf\a2Kf\1a\a8p\8bK\c2\a3Ql\c7\19\e8\92\d1$\06\99\d6\855\0e\f4p\a0j\10\16\c1\a4\19\08l7\1eLwH\'\b5\bc\b04\b3\0c\1c9J\aa\d8NO\ca\9c[\f3o.h\ee\82\8ftoc\a5x\14x\c8\84\08\02\c7\8c\fa\ff\be\90\eblP\a4\f7\a3\f9\be\f2xq\c6\00\00\00\00\00\00\00\00\00\00\00\00")
@@ -1797,11 +1798,6 @@
   (local $33 i32)
   (local $34 i32)
   (local $35 i32)
-  (local $36 i32)
-  (local $37 i32)
-  (local $38 i32)
-  (local $39 i32)
-  (local $40 i32)
   global.get $assembly/index/H0
   global.set $assembly/index/a
   global.get $assembly/index/H1
@@ -1822,127 +1818,131 @@
   global.set $assembly/index/i
   loop $for-loop|0
    global.get $assembly/index/i
-   i32.const 16
-   i32.lt_u
-   if
-    local.get $0
-    local.set $13
-    global.get $assembly/index/i
-    local.set $14
-    block $assembly/index/load32be|inlined.0 (result i32)
-     local.get $1
-     local.set $2
-     global.get $assembly/index/i
-     local.set $3
-     local.get $3
-     i32.const 2
-     i32.shl
-     local.set $4
-     block $assembly/index/load8|inlined.0 (result i32)
-      local.get $2
-      local.set $5
-      local.get $4
-      i32.const 0
-      i32.add
-      local.set $6
-      local.get $5
-      local.get $6
-      i32.add
-      i32.load8_u
-      br $assembly/index/load8|inlined.0
-     end
-     i32.const 255
-     i32.and
-     i32.const 24
-     i32.shl
-     block $assembly/index/load8|inlined.1 (result i32)
-      local.get $2
-      local.set $7
-      local.get $4
-      i32.const 1
-      i32.add
-      local.set $8
-      local.get $7
-      local.get $8
-      i32.add
-      i32.load8_u
-      br $assembly/index/load8|inlined.1
-     end
-     i32.const 255
-     i32.and
-     i32.const 16
-     i32.shl
-     i32.or
-     block $assembly/index/load8|inlined.2 (result i32)
-      local.get $2
-      local.set $9
-      local.get $4
-      i32.const 2
-      i32.add
-      local.set $10
-      local.get $9
-      local.get $10
-      i32.add
-      i32.load8_u
-      br $assembly/index/load8|inlined.2
-     end
-     i32.const 255
-     i32.and
-     i32.const 8
-     i32.shl
-     i32.or
-     block $assembly/index/load8|inlined.3 (result i32)
-      local.get $2
-      local.set $11
-      local.get $4
-      i32.const 3
-      i32.add
-      local.set $12
-      local.get $11
-      local.get $12
-      i32.add
-      i32.load8_u
-      br $assembly/index/load8|inlined.3
-     end
-     i32.const 255
-     i32.and
-     i32.const 0
-     i32.shl
-     i32.or
-     br $assembly/index/load32be|inlined.0
-    end
-    local.set $15
-    local.get $13
-    local.get $14
-    i32.const 2
-    i32.shl
-    i32.add
-    local.get $15
-    i32.store
-    global.get $assembly/index/i
-    i32.const 1
-    i32.add
-    global.set $assembly/index/i
-    br $for-loop|0
-   end
-  end
-  i32.const 16
-  global.set $assembly/index/i
-  loop $for-loop|1
-   global.get $assembly/index/i
    i32.const 64
    i32.lt_u
    if
-    local.get $0
-    local.set $26
     global.get $assembly/index/i
-    local.set $27
-    block $assembly/index/SIG1|inlined.0 (result i32)
-     block $assembly/index/load32|inlined.0 (result i32)
+    i32.const 16
+    i32.lt_u
+    if (result i32)
+     block $assembly/index/load32be|inlined.0 (result i32)
+      local.get $1
+      local.set $2
+      global.get $assembly/index/i
+      local.set $3
+      local.get $3
+      i32.const 2
+      i32.shl
+      local.set $4
+      block $assembly/index/load8|inlined.0 (result i32)
+       local.get $2
+       local.set $5
+       local.get $4
+       i32.const 0
+       i32.add
+       local.set $6
+       local.get $5
+       local.get $6
+       i32.add
+       i32.load8_u
+       br $assembly/index/load8|inlined.0
+      end
+      i32.const 255
+      i32.and
+      i32.const 24
+      i32.shl
+      block $assembly/index/load8|inlined.1 (result i32)
+       local.get $2
+       local.set $7
+       local.get $4
+       i32.const 1
+       i32.add
+       local.set $8
+       local.get $7
+       local.get $8
+       i32.add
+       i32.load8_u
+       br $assembly/index/load8|inlined.1
+      end
+      i32.const 255
+      i32.and
+      i32.const 16
+      i32.shl
+      i32.or
+      block $assembly/index/load8|inlined.2 (result i32)
+       local.get $2
+       local.set $9
+       local.get $4
+       i32.const 2
+       i32.add
+       local.set $10
+       local.get $9
+       local.get $10
+       i32.add
+       i32.load8_u
+       br $assembly/index/load8|inlined.2
+      end
+      i32.const 255
+      i32.and
+      i32.const 8
+      i32.shl
+      i32.or
+      block $assembly/index/load8|inlined.3 (result i32)
+       local.get $2
+       local.set $11
+       local.get $4
+       i32.const 3
+       i32.add
+       local.set $12
+       local.get $11
+       local.get $12
+       i32.add
+       i32.load8_u
+       br $assembly/index/load8|inlined.3
+      end
+      i32.const 255
+      i32.and
+      i32.const 0
+      i32.shl
+      i32.or
+      br $assembly/index/load32be|inlined.0
+     end
+    else
+     block $assembly/index/SIG1|inlined.0 (result i32)
+      block $assembly/index/load32|inlined.0 (result i32)
+       local.get $0
+       local.set $13
+       global.get $assembly/index/i
+       i32.const 2
+       i32.sub
+       local.set $14
+       local.get $13
+       local.get $14
+       i32.const 2
+       i32.shl
+       i32.add
+       i32.load
+       br $assembly/index/load32|inlined.0
+      end
+      local.set $15
+      local.get $15
+      i32.const 17
+      i32.rotr
+      local.get $15
+      i32.const 19
+      i32.rotr
+      i32.xor
+      local.get $15
+      i32.const 10
+      i32.shr_u
+      i32.xor
+      br $assembly/index/SIG1|inlined.0
+     end
+     block $assembly/index/load32|inlined.1 (result i32)
       local.get $0
       local.set $16
       global.get $assembly/index/i
-      i32.const 2
+      i32.const 7
       i32.sub
       local.set $17
       local.get $16
@@ -1951,44 +1951,45 @@
       i32.shl
       i32.add
       i32.load
-      br $assembly/index/load32|inlined.0
+      br $assembly/index/load32|inlined.1
      end
-     local.set $18
-     local.get $18
-     i32.const 17
-     i32.rotr
-     local.get $18
-     i32.const 19
-     i32.rotr
-     i32.xor
-     local.get $18
-     i32.const 10
-     i32.shr_u
-     i32.xor
-     br $assembly/index/SIG1|inlined.0
-    end
-    block $assembly/index/load32|inlined.1 (result i32)
-     local.get $0
-     local.set $19
-     global.get $assembly/index/i
-     i32.const 7
-     i32.sub
-     local.set $20
-     local.get $19
-     local.get $20
-     i32.const 2
-     i32.shl
      i32.add
-     i32.load
-     br $assembly/index/load32|inlined.1
-    end
-    i32.add
-    block $assembly/index/SIG0|inlined.0 (result i32)
-     block $assembly/index/load32|inlined.2 (result i32)
+     block $assembly/index/SIG0|inlined.0 (result i32)
+      block $assembly/index/load32|inlined.2 (result i32)
+       local.get $0
+       local.set $18
+       global.get $assembly/index/i
+       i32.const 15
+       i32.sub
+       local.set $19
+       local.get $18
+       local.get $19
+       i32.const 2
+       i32.shl
+       i32.add
+       i32.load
+       br $assembly/index/load32|inlined.2
+      end
+      local.set $20
+      local.get $20
+      i32.const 7
+      i32.rotr
+      local.get $20
+      i32.const 18
+      i32.rotr
+      i32.xor
+      local.get $20
+      i32.const 3
+      i32.shr_u
+      i32.xor
+      br $assembly/index/SIG0|inlined.0
+     end
+     i32.add
+     block $assembly/index/load32|inlined.3 (result i32)
       local.get $0
       local.set $21
       global.get $assembly/index/i
-      i32.const 15
+      i32.const 16
       i32.sub
       local.set $22
       local.get $21
@@ -1997,73 +1998,36 @@
       i32.shl
       i32.add
       i32.load
-      br $assembly/index/load32|inlined.2
+      br $assembly/index/load32|inlined.3
      end
-     local.set $23
-     local.get $23
-     i32.const 7
-     i32.rotr
-     local.get $23
-     i32.const 18
-     i32.rotr
-     i32.xor
-     local.get $23
-     i32.const 3
-     i32.shr_u
-     i32.xor
-     br $assembly/index/SIG0|inlined.0
-    end
-    i32.add
-    block $assembly/index/load32|inlined.3 (result i32)
-     local.get $0
-     local.set $24
-     global.get $assembly/index/i
-     i32.const 16
-     i32.sub
-     local.set $25
-     local.get $24
-     local.get $25
-     i32.const 2
-     i32.shl
      i32.add
-     i32.load
-     br $assembly/index/load32|inlined.3
     end
-    i32.add
-    local.set $28
-    local.get $26
-    local.get $27
+    global.set $assembly/index/tmpW
+    local.get $0
+    local.set $23
+    global.get $assembly/index/i
+    local.set $24
+    global.get $assembly/index/tmpW
+    local.set $25
+    local.get $23
+    local.get $24
     i32.const 2
     i32.shl
     i32.add
-    local.get $28
+    local.get $25
     i32.store
-    global.get $assembly/index/i
-    i32.const 1
-    i32.add
-    global.set $assembly/index/i
-    br $for-loop|1
-   end
-  end
-  i32.const 0
-  global.set $assembly/index/i
-  loop $for-loop|2
-   global.get $assembly/index/i
-   i32.const 64
-   i32.lt_u
-   if
     global.get $assembly/index/h
     block $assembly/index/EP1|inlined.0 (result i32)
      global.get $assembly/index/e
-     local.set $29
-     local.get $29
+     local.set $26
+     local.get $26
      i32.const 6
      i32.rotr
-     local.get $29
+     local.get $26
      i32.const 11
      i32.rotr
      i32.xor
-     local.get $29
+     local.get $26
      i32.const 25
      i32.rotr
      i32.xor
@@ -2072,18 +2036,18 @@
     i32.add
     block $assembly/index/CH|inlined.0 (result i32)
      global.get $assembly/index/e
-     local.set $30
+     local.set $27
      global.get $assembly/index/f
-     local.set $31
+     local.set $28
      global.get $assembly/index/g
-     local.set $32
-     local.get $30
-     local.get $31
+     local.set $29
+     local.get $27
+     local.get $28
      i32.and
-     local.get $30
+     local.get $27
      i32.const -1
      i32.xor
-     local.get $32
+     local.get $29
      i32.and
      i32.xor
      br $assembly/index/CH|inlined.0
@@ -2091,11 +2055,11 @@
     i32.add
     block $assembly/index/load32|inlined.4 (result i32)
      global.get $assembly/index/kPtr
-     local.set $33
+     local.set $30
      global.get $assembly/index/i
-     local.set $34
-     local.get $33
-     local.get $34
+     local.set $31
+     local.get $30
+     local.get $31
      i32.const 2
      i32.shl
      i32.add
@@ -2103,32 +2067,20 @@
      br $assembly/index/load32|inlined.4
     end
     i32.add
-    block $assembly/index/load32|inlined.5 (result i32)
-     local.get $0
-     local.set $35
-     global.get $assembly/index/i
-     local.set $36
-     local.get $35
-     local.get $36
-     i32.const 2
-     i32.shl
-     i32.add
-     i32.load
-     br $assembly/index/load32|inlined.5
-    end
+    global.get $assembly/index/tmpW
     i32.add
     global.set $assembly/index/t1
     block $assembly/index/EP0|inlined.0 (result i32)
      global.get $assembly/index/a
-     local.set $37
-     local.get $37
+     local.set $32
+     local.get $32
      i32.const 2
      i32.rotr
-     local.get $37
+     local.get $32
      i32.const 13
      i32.rotr
      i32.xor
-     local.get $37
+     local.get $32
      i32.const 22
      i32.rotr
      i32.xor
@@ -2136,20 +2088,20 @@
     end
     block $assembly/index/MAJ|inlined.0 (result i32)
      global.get $assembly/index/a
-     local.set $38
+     local.set $33
      global.get $assembly/index/b
-     local.set $39
+     local.set $34
      global.get $assembly/index/c
-     local.set $40
-     local.get $38
-     local.get $39
+     local.set $35
+     local.get $33
+     local.get $34
      i32.and
-     local.get $38
-     local.get $40
+     local.get $33
+     local.get $35
      i32.and
      i32.xor
-     local.get $39
-     local.get $40
+     local.get $34
+     local.get $35
      i32.and
      i32.xor
      br $assembly/index/MAJ|inlined.0
@@ -2180,7 +2132,7 @@
     i32.const 1
     i32.add
     global.set $assembly/index/i
-    br $for-loop|2
+    br $for-loop|0
    end
   end
   global.get $assembly/index/H0
@@ -2759,7 +2711,7 @@
      br $assembly/index/CH|inlined.1
     end
     i32.add
-    block $assembly/index/load32|inlined.6 (result i32)
+    block $assembly/index/load32|inlined.5 (result i32)
      local.get $0
      local.set $5
      global.get $assembly/index/i
@@ -2770,7 +2722,7 @@
      i32.shl
      i32.add
      i32.load
-     br $assembly/index/load32|inlined.6
+     br $assembly/index/load32|inlined.5
     end
     i32.add
     global.set $assembly/index/t1