… packages/as-sha256/build/optimized.wasm
… packages/as-sha256/build/untouched.wasm
… packages/as-sha256/src/wasmCode.ts