**Motivation**

- Improve as-sha256

**Description**

- Chain w computation in the same for loop of hashBlocks()
- Use UintArray.slice()

Closes #357