@tuyennhv 2024-03-25
benchmark is consistently worse in CI, however on feat4 mainnet node and my local env it's better

master
```
 ✓ digestTwoHashObjects 50023 times                                    20.73418 ops/s    48.22953 ms/op        -       1233 runs   60.0 s
 ✓ digest64 50023 times                                                19.74772 ops/s    50.63876 ms/op        -       1175 runs   60.0 s
 ✓ digest 50023 times                                                  19.73773 ops/s    50.66439 ms/op        -       1174 runs   60.0 s
```

this branch
```
✓ digestTwoHashObjects 50023 times                                    22.70416 ops/s    44.04479 ms/op        -       1351 runs   60.0 s
✓ digest64 50023 times                                                21.21962 ops/s    47.12619 ms/op        -       1263 runs   60.0 s
✓ digest 50023 times                                                  21.01592 ops/s    47.58299 ms/op        -       1250 runs   60.0 s
```